refactor predictions into different method
this commit does two important things (with some hopefully benign side
effects). firstly, i have refactored the prediction code from the
VariationalMergingModel.call method into a method called
VariationalMergingModel.prediction_mean_stddev which uses scipy to
compute the first and second moments of model predictions. the caveat with this is
that we no longer have access to a differentiable first moment function.
this should not be an issue for any foreseeable use case. secondly, this
commit adapts the DataManager class to use the new moment function and
output uncertainties in the predictions.